% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaclean-package.R
\docType{package}
\name{cleaclean-package}
\alias{cleaclean}
\alias{cleaclean-package}
\title{cleaclean: Clean Constituency-Level Election Results}
\description{
\code{cleaclean} provides two cleaned election datasets derived from the
Constituency-Level Elections Archive. The package exports datasets only;
the repository contains the audited maintainer workflow used to build them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jackobailey/clean_clea}
  \item Report bugs at \url{https://github.com/jackobailey/clean_clea/issues}
}

}
\author{
\strong{Maintainer}: Jack Bailey \email{37049065+jackobailey@users.noreply.github.com}

Other contributors:
\itemize{
  \item Chris Hanretty \email{chris.hanretty@gmail.com} [contributor]
}

}
\keyword{internal}
